settings and plugin rename
