+++ services/web/server/src/simcore_service_webserver/application_settings.py
@@ -33,7 +33,7 @@ from .director.settings import DirectorSettings
 from .director_v2_settings import DirectorV2Settings
 from .exporter.settings import ExporterSettings
 from .garbage_collector_settings import GarbageCollectorSettings
-from .invitations.invitations_settings import InvitationsSettings
+from .invitations.settings import InvitationsSettings
 from .login.settings import LoginSettings
 from .projects.projects_settings import ProjectsSettings
 from .resource_manager.settings import ResourceManagerSettings
